refactor: update package structure for better modularization and clarity
