+++ internal/external/openweather/provider.go
@@ -1,4 +1,4 @@
-package weather
+package openweather
 
 import (
 	"context"
