+++ internal/external/sendgrid/sendgrid.go
@@ -1,4 +1,4 @@
-package external
+package sendgrid
 
 import (
 	"fmt"
